requirements.txt
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
altgraph==0.17.2
anyio==4.2.0
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==44.0.1
cycler==0.12.1
disnake==2.9.0
distlib==0.3.8
exceptiongroup==1.2.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.44.3
frozenlist==1.4.0
fsspec==2025.3.2
future==0.18.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.0.2
gspread_asyncio==2.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-resources==6.1.1
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==4.9.4
macholib==1.15.2
MarkupSafe==2.1.5
matplotlib==3.8.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pdfminer.six==20231228
pdfplumber==0.11.5
Pillow==10.1.0
platformdirs==4.2.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.0
pyparsing==3.1.1
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.13.1
seaborn==0.13.2
selenium==4.28.1
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
StrEnum==0.4.15
sympy==1.13.3
tesseract==0.1.3
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
twill==3.2.2
typing_extensions==4.13.2
tzdata==2023.3
ultralytics==8.3.118
ultralytics-thop==2.0.14
urllib3==2.2.0
virtualenv==20.26.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
opencv-python-headless
base64
gunicorn